Make the built-in visitors private
