+++ serde/src/de/mod.rs
@@ -107,16 +107,19 @@ use core::marker::PhantomData;
 
 ///////////////////////////////////////////////////////////////////////////////
 
-#[doc(hidden)]
-pub mod impls;
 pub mod value;
+
 mod from_primitive;
+mod ignored_any;
+mod impls;
 
 // Helpers used by generated code. Not public API.
 #[doc(hidden)]
 pub mod private;
 #[cfg(any(feature = "std", feature = "collections"))]
 mod content;
+
+pub use self::ignored_any::IgnoredAny;
 
 ///////////////////////////////////////////////////////////////////////////////
